Stubbed out some simple role READMEs
These roles are both very likely to change very much over coming commits. The logged_in role is liable to be better built with well-defined phases and proper handlers for logging out, to be defined as a dependency by other roles in the collection which may need it (like the operators role). The operators role is currently only half-implemented as the variable schema isn't set in stone yet, and is currently blocked by some fixes in oc-mirror's test case bundles and catalogs. Because of this, the readmes are very simple and basically just token contributions to get them into Ansible Galaxy.